auth.js に initFirebaseIfNeeded と isLoggedIn を追加
